python_flask_API/requirements.txt
Flask
pandas
plotly
mysql-connector-python
flasgger
PyYAML